BSERDEV-12904 Zero Downtime Backup

+++ bitbucket.diy-backup.common.sh
@@ -9,6 +9,9 @@ BITBUCKET_HTTP_AUTH="-u ${BITBUCKET_BACKUP_USER}:${BITBUCKET_BACKUP_PASS}"
 PRODUCT=Bitbucket
 
 function bitbucket_lock {
+    if [ "${BACKUP_ZERO_DOWNTIME}" = "true" ]; then
+        return
+    fi
     BITBUCKET_LOCK_RESULT=`curl ${CURL_OPTIONS} ${BITBUCKET_HTTP_AUTH} -X POST -H "Content-type: application/json" "${BITBUCKET_URL}/mvc/maintenance/lock"`
     if [ -z "${BITBUCKET_LOCK_RESULT}" ]; then
         bail "Locking this Bitbucket instance failed"
@@ -23,6 +26,10 @@ function bitbucket_lock {
 }
 
 function bitbucket_backup_start {
+    if [ "${BACKUP_ZERO_DOWNTIME}" = "true" ]; then
+        return
+    fi
+
     BITBUCKET_BACKUP_RESULT=`curl ${CURL_OPTIONS} ${BITBUCKET_HTTP_AUTH} -X POST -H "X-Atlassian-Maintenance-Token: ${BITBUCKET_LOCK_TOKEN}" -H "Accept: application/json" -H "Content-type: application/json" "${BITBUCKET_URL}/mvc/admin/backups?external=true"`
     if [ -z "${BITBUCKET_BACKUP_RESULT}" ]; then
         bail "Entering backup mode failed"
@@ -37,6 +44,10 @@ function bitbucket_backup_start {
 }
 
 function bitbucket_backup_wait {
+    if [ "${BACKUP_ZERO_DOWNTIME}" = "true" ]; then
+        return
+    fi
+
     BITBUCKET_PROGRESS_DB_STATE="AVAILABLE"
     BITBUCKET_PROGRESS_SCM_STATE="AVAILABLE"
 
@@ -66,6 +77,10 @@ function bitbucket_backup_wait {
 }
 
 function bitbucket_backup_progress {
+    if [ "${BACKUP_ZERO_DOWNTIME}" = "true" ]; then
+        return
+    fi
+
     BITBUCKET_REPORT_RESULT=`curl ${CURL_OPTIONS} ${BITBUCKET_HTTP_AUTH} -X POST -H "Accept: application/json" -H "Content-type: application/json" "${BITBUCKET_URL}/mvc/admin/backups/progress/client?token=${BITBUCKET_LOCK_TOKEN}&percentage=$1"`
     if [ $? != 0 ]; then
         bail "Unable to update backup progress"
@@ -75,6 +90,10 @@ function bitbucket_backup_progress {
 }
 
 function bitbucket_unlock {
+    if [ "${BACKUP_ZERO_DOWNTIME}" = "true" ]; then
+        return
+    fi
+
     BITBUCKET_UNLOCK_RESULT=`curl ${CURL_OPTIONS} ${BITBUCKET_HTTP_AUTH} -X DELETE -H "Accept: application/json" -H "Content-type: application/json" "${BITBUCKET_URL}/mvc/maintenance/lock?token=${BITBUCKET_LOCK_TOKEN}"`
     if [ $? != 0 ]; then
         bail "Unable to unlock instance with lock ${BITBUCKET_LOCK_TOKEN}"
